# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProposAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProposalStatus =
  (typeof ProposalStatus)[keyof typeof ProposalStatus];

export const ProposalStatus = {
  draft: "draft",
  sent: "sent",
  accepted: "accepted",
  declined: "declined",
} as const;

export interface Proposal {
  id: number;
  userId: string;
  clientName: string;
  clientEmail?: string | null;
  projectTitle: string;
  industry?: string | null;
  projectDescription?: string | null;
  budget?: string | null;
  timeline?: string | null;
  status: ProposalStatus;
  executiveSummary?: string | null;
  understanding?: string | null;
  approach?: string | null;
  timelinePlan?: string | null;
  investment?: string | null;
  whyUs?: string | null;
  nextSteps?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProposalList {
  proposals: Proposal[];
  total: number;
}

export interface CreateProposalBody {
  clientName: string;
  clientEmail?: string | null;
  projectTitle: string;
  industry?: string | null;
  projectDescription?: string | null;
  budget?: string | null;
  timeline?: string | null;
}

export interface UpdateProposalBody {
  clientName?: string;
  clientEmail?: string | null;
  projectTitle?: string;
  industry?: string | null;
  projectDescription?: string | null;
  budget?: string | null;
  timeline?: string | null;
  executiveSummary?: string | null;
  understanding?: string | null;
  approach?: string | null;
  timelinePlan?: string | null;
  investment?: string | null;
  whyUs?: string | null;
  nextSteps?: string | null;
}

export interface UpdateProposalStatusBody {
  status: ProposalStatus;
}

export type GenerateProposalContentBodySection =
  (typeof GenerateProposalContentBodySection)[keyof typeof GenerateProposalContentBodySection];

export const GenerateProposalContentBodySection = {
  executiveSummary: "executiveSummary",
  understanding: "understanding",
  approach: "approach",
  timelinePlan: "timelinePlan",
  investment: "investment",
  whyUs: "whyUs",
  nextSteps: "nextSteps",
  all: "all",
} as const;

export interface GenerateProposalContentBody {
  section: GenerateProposalContentBodySection;
  additionalContext?: string | null;
}

export interface DashboardStats {
  totalProposals: number;
  draftCount: number;
  sentCount: number;
  acceptedCount: number;
  declinedCount: number;
  winRate: number;
}

export type ListProposalsParams = {
  status?: ListProposalsStatus;
  limit?: number;
  offset?: number;
};

export type ListProposalsStatus =
  (typeof ListProposalsStatus)[keyof typeof ListProposalsStatus];

export const ListProposalsStatus = {
  draft: "draft",
  sent: "sent",
  accepted: "accepted",
  declined: "declined",
} as const;

export type GetRecentProposalsParams = {
  limit?: number;
};