# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProposAI API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all proposals for the authenticated user
 */
export const listProposalsQueryLimitDefault = 20;
export const listProposalsQueryOffsetDefault = 0;

export const ListProposalsQueryParams = zod.object({
  status: zod.enum(["draft", "sent", "accepted", "declined"]).optional(),
  limit: zod.coerce.number().default(listProposalsQueryLimitDefault),
  offset: zod.coerce.number().default(listProposalsQueryOffsetDefault),
});

export const ListProposalsResponse = zod.object({
  proposals: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.string(),
      clientName: zod.string(),
      clientEmail: zod.string().nullish(),
      projectTitle: zod.string(),
      industry: zod.string().nullish(),
      projectDescription: zod.string().nullish(),
      budget: zod.string().nullish(),
      timeline: zod.string().nullish(),
      status: zod.enum(["draft", "sent", "accepted", "declined"]),
      executiveSummary: zod.string().nullish(),
      understanding: zod.string().nullish(),
      approach: zod.string().nullish(),
      timelinePlan: zod.string().nullish(),
      investment: zod.string().nullish(),
      whyUs: zod.string().nullish(),
      nextSteps: zod.string().nullish(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Create a new proposal
 */
export const CreateProposalBody = zod.object({
  clientName: zod.string(),
  clientEmail: zod.string().nullish(),
  projectTitle: zod.string(),
  industry: zod.string().nullish(),
  projectDescription: zod.string().nullish(),
  budget: zod.string().nullish(),
  timeline: zod.string().nullish(),
});

/**
 * @summary Get a proposal by ID
 */
export const GetProposalParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProposalResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  clientName: zod.string(),
  clientEmail: zod.string().nullish(),
  projectTitle: zod.string(),
  industry: zod.string().nullish(),
  projectDescription: zod.string().nullish(),
  budget: zod.string().nullish(),
  timeline: zod.string().nullish(),
  status: zod.enum(["draft", "sent", "accepted", "declined"]),
  executiveSummary: zod.string().nullish(),
  understanding: zod.string().nullish(),
  approach: zod.string().nullish(),
  timelinePlan: zod.string().nullish(),
  investment: zod.string().nullish(),
  whyUs: zod.string().nullish(),
  nextSteps: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a proposal
 */
export const UpdateProposalParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProposalBody = zod.object({
  clientName: zod.string().optional(),
  clientEmail: zod.string().nullish(),
  projectTitle: zod.string().optional(),
  industry: zod.string().nullish(),
  projectDescription: zod.string().nullish(),
  budget: zod.string().nullish(),
  timeline: zod.string().nullish(),
  executiveSummary: zod.string().nullish(),
  understanding: zod.string().nullish(),
  approach: zod.string().nullish(),
  timelinePlan: zod.string().nullish(),
  investment: zod.string().nullish(),
  whyUs: zod.string().nullish(),
  nextSteps: zod.string().nullish(),
});

export const UpdateProposalResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  clientName: zod.string(),
  clientEmail: zod.string().nullish(),
  projectTitle: zod.string(),
  industry: zod.string().nullish(),
  projectDescription: zod.string().nullish(),
  budget: zod.string().nullish(),
  timeline: zod.string().nullish(),
  status: zod.enum(["draft", "sent", "accepted", "declined"]),
  executiveSummary: zod.string().nullish(),
  understanding: zod.string().nullish(),
  approach: zod.string().nullish(),
  timelinePlan: zod.string().nullish(),
  investment: zod.string().nullish(),
  whyUs: zod.string().nullish(),
  nextSteps: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a proposal
 */
export const DeleteProposalParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Generate AI content for a proposal section
 */
export const GenerateProposalContentParams = zod.object({
  id: zod.coerce.number(),
});

export const GenerateProposalContentBody = zod.object({
  section: zod.enum([
    "executiveSummary",
    "understanding",
    "approach",
    "timelinePlan",
    "investment",
    "whyUs",
    "nextSteps",
    "all",
  ]),
  additionalContext: zod.string().nullish(),
});

/**
 * @summary Duplicate an existing proposal
 */
export const DuplicateProposalParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Update proposal status
 */
export const UpdateProposalStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProposalStatusBody = zod.object({
  status: zod.enum(["draft", "sent", "accepted", "declined"]),
});

export const UpdateProposalStatusResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  clientName: zod.string(),
  clientEmail: zod.string().nullish(),
  projectTitle: zod.string(),
  industry: zod.string().nullish(),
  projectDescription: zod.string().nullish(),
  budget: zod.string().nullish(),
  timeline: zod.string().nullish(),
  status: zod.enum(["draft", "sent", "accepted", "declined"]),
  executiveSummary: zod.string().nullish(),
  understanding: zod.string().nullish(),
  approach: zod.string().nullish(),
  timelinePlan: zod.string().nullish(),
  investment: zod.string().nullish(),
  whyUs: zod.string().nullish(),
  nextSteps: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Get dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalProposals: zod.number(),
  draftCount: zod.number(),
  sentCount: zod.number(),
  acceptedCount: zod.number(),
  declinedCount: zod.number(),
  winRate: zod.number(),
});

/**
 * @summary Get recent proposals for dashboard feed
 */
export const getRecentProposalsQueryLimitDefault = 5;

export const GetRecentProposalsQueryParams = zod.object({
  limit: zod.coerce.number().default(getRecentProposalsQueryLimitDefault),
});

export const GetRecentProposalsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  clientName: zod.string(),
  clientEmail: zod.string().nullish(),
  projectTitle: zod.string(),
  industry: zod.string().nullish(),
  projectDescription: zod.string().nullish(),
  budget: zod.string().nullish(),
  timeline: zod.string().nullish(),
  status: zod.enum(["draft", "sent", "accepted", "declined"]),
  executiveSummary: zod.string().nullish(),
  understanding: zod.string().nullish(),
  approach: zod.string().nullish(),
  timelinePlan: zod.string().nullish(),
  investment: zod.string().nullish(),
  whyUs: zod.string().nullish(),
  nextSteps: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetRecentProposalsResponse = zod.array(
  GetRecentProposalsResponseItem,
);