# lib/api-zod/src/generated/types/createProposalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProposAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProposalBody {
  clientName: string;
  clientEmail?: string | null;
  projectTitle: string;
  industry?: string | null;
  projectDescription?: string | null;
  budget?: string | null;
  timeline?: string | null;
}