# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProposAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalProposals: number;
  draftCount: number;
  sentCount: number;
  acceptedCount: number;
  declinedCount: number;
  winRate: number;
}