# lib/api-zod/src/generated/types/generateProposalContentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProposAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateProposalContentBodySection } from "./generateProposalContentBodySection";

export interface GenerateProposalContentBody {
  section: GenerateProposalContentBodySection;
  additionalContext?: string | null;
}